chore: move phone numer validation to unique convex validation
